Wed, Mar 27, 2024, 11:57 AM -07:00

+++ filter-more/helpers.c
@@ -126,9 +126,9 @@ void edges(int height, int width, RGBTRIPLE image[height][width])
             // array of sobel values of RGB
             int SobelRGB[3];
             // calculate square root of Gx^2+Gy^2
-            SobelRGB[0] = round(sqrt(pow(GxRed, 2) + pow(GyRed, 2)));
-            SobelRGB[1] = round(sqrt(pow(GxGreen, 2) + pow(GyGreen, 2)));
-            SobelRGB[2] = round(sqrt(pow(GxBlue, 2) + pow(GyBlue, 2)));
+            SobelRGB[0] = round(sqrt(pow((double)GxRed, 2) + pow((double)GyRed, 2)));
+            SobelRGB[1] = round(sqrt(pow((double)GxGreen, 2) + pow((double)GyGreen, 2)));
+            SobelRGB[2] = round(sqrt(pow((double)GxBlue, 2) + pow((double)GyBlue, 2)));
 
             // deal with values exceeding 255
             for (int k = 0; k < 3; k++)
